refactor: capture currentUrl and wire up to redux
